Move code generation to `MissionModelGenerator`
Also rename activity mapper "instantiators" to "stylers".
These classes simply exist to "style" an activity mapper depending on the activity "style".
