+++ merlin-framework-processor/src/main/java/gov/nasa/jpl/aerie/merlin/processor/MissionModelProcessor.java
@@ -1,34 +1,8 @@
 package gov.nasa.jpl.aerie.merlin.processor;
 
-import com.squareup.javapoet.AnnotationSpec;
-import com.squareup.javapoet.ClassName;
-import com.squareup.javapoet.CodeBlock;
-import com.squareup.javapoet.FieldSpec;
-import com.squareup.javapoet.JavaFile;
-import com.squareup.javapoet.MethodSpec;
-import com.squareup.javapoet.ParameterSpec;
-import com.squareup.javapoet.ParameterizedTypeName;
-import com.squareup.javapoet.TypeName;
-import com.squareup.javapoet.TypeSpec;
-import com.squareup.javapoet.WildcardTypeName;
 import gov.nasa.jpl.aerie.merlin.framework.annotations.ActivityType;
 import gov.nasa.jpl.aerie.merlin.framework.annotations.MissionModel;
-import gov.nasa.jpl.aerie.merlin.processor.instantiators.ActivityMapperInstantiator;
-import gov.nasa.jpl.aerie.merlin.processor.instantiators.AllStaticallyDefinedInstantiator;
-import gov.nasa.jpl.aerie.merlin.processor.instantiators.NoneDefinedInstantiator;
-import gov.nasa.jpl.aerie.merlin.processor.instantiators.AllDefinedInstantiator;
-import gov.nasa.jpl.aerie.merlin.processor.instantiators.SomeStaticallyDefinedInstantiator;
-import gov.nasa.jpl.aerie.merlin.processor.metamodel.ActivityMapperRecord;
-import gov.nasa.jpl.aerie.merlin.processor.metamodel.ActivityParameterRecord;
-import gov.nasa.jpl.aerie.merlin.processor.metamodel.ActivityTypeRecord;
-import gov.nasa.jpl.aerie.merlin.processor.metamodel.ActivityValidationRecord;
-import gov.nasa.jpl.aerie.merlin.processor.metamodel.MissionModelRecord;
-import gov.nasa.jpl.aerie.merlin.processor.metamodel.TypeRule;
-import gov.nasa.jpl.aerie.merlin.protocol.model.MerlinPlugin;
-import gov.nasa.jpl.aerie.merlin.protocol.model.MissionModelFactory;
-import gov.nasa.jpl.aerie.merlin.protocol.types.Parameter;
-
-import org.apache.commons.lang3.tuple.Pair;
+import gov.nasa.jpl.aerie.merlin.processor.generator.MissionModelGenerator;
 
 import javax.annotation.processing.Completion;
 import javax.annotation.processing.Filer;
@@ -38,31 +12,25 @@ import javax.annotation.processing.Processor;
 import javax.annotation.processing.RoundEnvironment;
 import javax.lang.model.SourceVersion;
 import javax.lang.model.element.AnnotationMirror;
-import javax.lang.model.element.AnnotationValue;
 import javax.lang.model.element.Element;
-import javax.lang.model.element.ElementKind;
 import javax.lang.model.element.ExecutableElement;
-import javax.lang.model.element.Modifier;
 import javax.lang.model.element.PackageElement;
 import javax.lang.model.element.TypeElement;
-import javax.lang.model.type.DeclaredType;
 import javax.lang.model.util.Elements;
 import javax.lang.model.util.Types;
 import javax.tools.Diagnostic;
-import java.lang.annotation.Annotation;
-import java.lang.annotation.Repeatable;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Map;
-import java.util.Objects;
-import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
 
+/**
+ * Top-level annotation processor for mission models.
+ * Involved parsing and code generation is handled by the {@link MissionModelParser} and {@link MissionModelGenerator} respectively.
+ */
 public final class MissionModelProcessor implements Processor {
 
   private final Set<Element> foundActivityTypes = new HashSet<>();
@@ -105,23 +73,25 @@ public final class MissionModelProcessor implements Processor {
     // Accumulate any information added in this round.
     this.foundActivityTypes.addAll(roundEnv.getElementsAnnotatedWith(ActivityType.class));
 
+    final var missionModelParser = new MissionModelParser(elementUtils, typeUtils);
+    final var missionModelGen = new MissionModelGenerator(elementUtils, typeUtils, messager);
+
     // Iterate over all elements annotated with @MissionModel
     for (final var element : roundEnv.getElementsAnnotatedWith(MissionModel.class)) {
       final var packageElement = (PackageElement) element;
-      final var missionModelParser = new MissionModelParser(elementUtils, typeUtils);
-
       try {
         final var missionModelRecord = missionModelParser.parseMissionModel(packageElement);
 
-        final var generatedFiles = new ArrayList<JavaFile>();
-        generatedFiles.add(generateMerlinPlugin(missionModelRecord));
-        generatedFiles.add(generateMissionModelFactory(missionModelRecord));
-        generatedFiles.add(generateActivityActions(missionModelRecord));
-        generatedFiles.add(generateActivityTypes(missionModelRecord));
+        final var generatedFiles = new ArrayList<>(List.of(
+            missionModelGen.generateMerlinPlugin(missionModelRecord),
+            missionModelGen.generateMissionModelFactory(missionModelRecord),
+            missionModelGen.generateActivityActions(missionModelRecord),
+            missionModelGen.generateActivityTypes(missionModelRecord)
+        ));
         for (final var activityRecord : missionModelRecord.activityTypes) {
           this.ownedActivityTypes.add(activityRecord.declaration);
           if (!activityRecord.mapper.isCustom) {
-            generateActivityMapper(missionModelRecord, activityRecord).ifPresent(generatedFiles::add);
+            missionModelGen.generateActivityMapper(missionModelRecord, activityRecord).ifPresent(generatedFiles::add);
           }
         }
 
@@ -182,555 +152,5 @@ public final class MissionModelProcessor implements Processor {
       final String userText)
   {
     return Collections::emptyIterator;
-  }
-
-  private static ActivityMapperInstantiator getMapperInstantiator(final ActivityDefaultsStyle style) {
-    return switch (style) {
-      case AllStaticallyDefined -> new AllStaticallyDefinedInstantiator();
-      case NoneDefined -> new NoneDefinedInstantiator();
-      case AllDefined -> new AllDefinedInstantiator();
-      case SomeStaticallyDefined -> new SomeStaticallyDefinedInstantiator();
-    };
-  }
-
-  private Optional<Map<String, CodeBlock>> buildParameterMapperBlocks(final MissionModelRecord missionModel, final ActivityTypeRecord activityType)
-  {
-    final var resolver = new Resolver(this.typeUtils, this.elementUtils, missionModel.typeRules);
-    var failed = false;
-    final var mapperBlocks = new HashMap<String, CodeBlock>();
-
-    for (final var parameter : activityType.parameters) {
-      final var mapperBlock = resolver.instantiateNullableMapperFor(parameter.type);
-      if (mapperBlock.isPresent()) {
-        mapperBlocks.put(parameter.name, mapperBlock.get());
-      } else {
-        failed = true;
-        messager.printMessage(
-            Diagnostic.Kind.ERROR,
-            "Failed to generate value mapper for parameter",
-            parameter.element);
-      }
-    }
-
-    return failed ? Optional.empty() : Optional.of(mapperBlocks);
-  }
-
-  private CodeBlock buildConfigurationMapperBlock(final MissionModelRecord missionModel, final TypeElement typeElem) {
-    final var resolver = new Resolver(this.typeUtils, this.elementUtils, missionModel.typeRules);
-    final var mapperBlock = resolver.instantiateMapperFor(typeElem.asType());
-    if (mapperBlock.isEmpty()) {
-      messager.printMessage(Diagnostic.Kind.ERROR, "Failed to generate value mapper for configuration", typeElem);
-    }
-    return mapperBlock.get();
-  }
-
-  private Optional<JavaFile> generateActivityMapper(final MissionModelRecord missionModel, final ActivityTypeRecord activityType)
-  {
-    final var maybeMapperBlocks = buildParameterMapperBlocks(missionModel, activityType);
-    if (maybeMapperBlocks.isEmpty()) return Optional.empty();
-    final var mapperBlocks = maybeMapperBlocks.get();
-
-    final var typeSpec =
-        TypeSpec
-            .classBuilder(activityType.mapper.name)
-            // The location of the missionModel package determines where to put this class.
-            .addOriginatingElement(missionModel.$package)
-            // The fields and methods of the activity determines the overall behavior of this class.
-            .addOriginatingElement(activityType.declaration)
-            // TODO: Add an originating element for each of the mapper rulesets associated with the mission model.
-            .addAnnotation(
-                AnnotationSpec
-                    .builder(javax.annotation.processing.Generated.class)
-                    .addMember("value", "$S", MissionModelProcessor.class.getCanonicalName())
-                    .build())
-            .addSuperinterface(
-                ParameterizedTypeName.get(
-                    ClassName.get(gov.nasa.jpl.aerie.merlin.protocol.model.TaskSpecType.class),
-                    ParameterizedTypeName.get(
-                        ClassName.get(gov.nasa.jpl.aerie.merlin.framework.RootModel.class),
-                        ClassName.get(missionModel.topLevelModel)),
-                    ClassName.get(activityType.declaration)))
-            .addModifiers(Modifier.PUBLIC, Modifier.FINAL)
-            .addFields(
-                activityType.parameters
-                    .stream()
-                    .map(parameter -> FieldSpec
-                        .builder(
-                            ParameterizedTypeName.get(
-                                ClassName.get(gov.nasa.jpl.aerie.merlin.framework.ValueMapper.class),
-                                TypeName.get(parameter.type).box()),
-                            "mapper_" + parameter.name)
-                        .addModifiers(Modifier.PRIVATE, Modifier.FINAL)
-                        .build())
-                    .collect(Collectors.toList()))
-            .addMethod(
-                MethodSpec
-                    .constructorBuilder()
-                    .addModifiers(Modifier.PUBLIC)
-                    /* Suppress unchecked warnings because the resolver has to
-                        put some big casting in for Class parameters
-                     */
-                    .addAnnotation(
-                        AnnotationSpec
-                            .builder(SuppressWarnings.class)
-                            .addMember("value", "$S", "unchecked")
-                            .build())
-                    .addCode(
-                        activityType.parameters
-                            .stream()
-                            .map(parameter -> CodeBlock
-                                .builder()
-                                .addStatement(
-                                    "this.mapper_$L =\n$L",
-                                    parameter.name,
-                                    mapperBlocks.get(parameter.name)))
-                            .reduce(CodeBlock.builder(), (x, y) -> x.add(y.build()))
-                            .build())
-                    .build())
-            .addMethod(
-                MethodSpec
-                    .methodBuilder("getName")
-                    .addModifiers(Modifier.PUBLIC)
-                    .addAnnotation(Override.class)
-                    .returns(String.class)
-                    .addStatement("return $S", activityType.name)
-                    .build())
-            .addMethod(getMapperInstantiator(activityType.activityDefaultsStyle).makeGetRequiredParametersMethod(activityType))
-            .addMethod(getMapperInstantiator(activityType.activityDefaultsStyle).makeGetParametersMethod(activityType))
-            .addMethod(getMapperInstantiator(activityType.activityDefaultsStyle).makeGetArgumentsMethod(activityType))
-            .addMethod(getMapperInstantiator(activityType.activityDefaultsStyle).makeInstantiateMethod(activityType))
-            .addMethod(
-                MethodSpec
-                    .methodBuilder("getValidationFailures")
-                    .addModifiers(Modifier.PUBLIC)
-                    .addAnnotation(Override.class)
-                    .returns(ParameterizedTypeName.get(
-                        java.util.List.class,
-                        String.class))
-                    .addParameter(
-                        TypeName.get(activityType.declaration.asType()),
-                        "activity",
-                        Modifier.FINAL)
-                    .addStatement(
-                        "final var $L = new $T()",
-                        "failures",
-                        ParameterizedTypeName.get(
-                            java.util.ArrayList.class,
-                            String.class))
-                    .addCode(
-                        activityType.validations
-                            .stream()
-                            .map(validation -> CodeBlock
-                                .builder()
-                                .addStatement(
-                                    "if (!$L.$L()) failures.add($S)",
-                                    "activity",
-                                    validation.methodName,
-                                    validation.failureMessage))
-                            .reduce(CodeBlock.builder(), (x, y) -> x.add(y.build()))
-                            .build())
-                    .addStatement(
-                        "return $L",
-                        "failures")
-                    .build())
-            .addMethod(
-                MethodSpec
-                    .methodBuilder("createTask")
-                    .addModifiers(Modifier.PUBLIC)
-                    .addAnnotation(Override.class)
-                    .returns(ClassName.get(gov.nasa.jpl.aerie.merlin.protocol.model.Task.class))
-                    .addParameter(
-                        ParameterizedTypeName.get(
-                            ClassName.get(gov.nasa.jpl.aerie.merlin.framework.RootModel.class),
-                            ClassName.get(missionModel.topLevelModel)),
-                        "model",
-                        Modifier.FINAL)
-                    .addParameter(
-                        TypeName.get(activityType.declaration.asType()),
-                        "activity",
-                        Modifier.FINAL)
-                    .addCode(
-                        activityType.effectModel
-                            .map(effectModel -> switch (effectModel.getRight()) {
-                              case Threaded -> CodeBlock
-                                  .builder()
-                                  .addStatement(
-                                      "return $T.threaded(() -> $L.$L($L.model())).create($L.executor())",
-                                      gov.nasa.jpl.aerie.merlin.framework.ModelActions.class,
-                                      "activity",
-                                      effectModel.getLeft(),
-                                      "model",
-                                      "model")
-                                  .build();
-
-                              case Replaying -> CodeBlock
-                                  .builder()
-                                  .addStatement(
-                                      "return $T.replaying(() -> $L.$L($L.model())).create($L.executor())",
-                                      gov.nasa.jpl.aerie.merlin.framework.ModelActions.class,
-                                      "activity",
-                                      effectModel.getLeft(),
-                                      "model",
-                                      "model")
-                                  .build();
-                            })
-                            .orElseGet(() -> CodeBlock
-                                .builder()
-                                .addStatement(
-                                    "return new $T($$ -> {})",
-                                    gov.nasa.jpl.aerie.merlin.framework.OneShotTask.class)
-                                .build()))
-                    .build())
-            .build();
-
-    return Optional.of(JavaFile
-                           .builder(activityType.mapper.name.packageName(), typeSpec)
-                           .skipJavaLangImports(true)
-                           .build());
-  }
-
-  private JavaFile generateActivityActions(final MissionModelRecord missionModel) {
-    final var typeName = missionModel.getActivityActionsName();
-
-    final var typeSpec =
-        TypeSpec
-            .classBuilder(typeName)
-            // The location of the mission model package determines where to put this class.
-            .addOriginatingElement(missionModel.$package)
-            // TODO: List found task spec types as dependencies of this generated file.
-            .addAnnotation(
-                AnnotationSpec
-                    .builder(javax.annotation.processing.Generated.class)
-                    .addMember("value", "$S", MissionModelProcessor.class.getCanonicalName())
-                    .build())
-            .addModifiers(Modifier.PUBLIC)
-            .superclass(gov.nasa.jpl.aerie.merlin.framework.ModelActions.class)
-            .addMethod(MethodSpec.constructorBuilder().addModifiers(Modifier.PRIVATE).build())
-            .addMethods(
-                missionModel.activityTypes
-                    .stream()
-                    .flatMap(entry -> List
-                        .of(
-                            MethodSpec
-                                .methodBuilder("spawn")
-                                .addModifiers(Modifier.PUBLIC, Modifier.STATIC)
-                                .returns(String.class)
-                                .addParameter(
-                                    ClassName.get(entry.declaration),
-                                    "activity",
-                                    Modifier.FINAL)
-                                .addStatement(
-                                    "final var $L = new $T()",
-                                    "mapper",
-                                    entry.mapper.name)
-                                .addStatement(
-                                    "return $T.spawn($L.getName(), $L.getArguments($L))",
-                                    gov.nasa.jpl.aerie.merlin.framework.ModelActions.class,
-                                    "mapper",
-                                    "mapper",
-                                    "activity")
-                                .build(),
-                            MethodSpec
-                                .methodBuilder("defer")
-                                .addModifiers(Modifier.PUBLIC, Modifier.STATIC)
-                                .returns(String.class)
-                                .addParameter(
-                                    ParameterSpec
-                                        .builder(
-                                            gov.nasa.jpl.aerie.merlin.protocol.types.Duration.class,
-                                            "duration")
-                                        .addModifiers(Modifier.FINAL)
-                                        .build())
-                                .addParameter(
-                                    ClassName.get(entry.declaration),
-                                    "activity",
-                                    Modifier.FINAL)
-                                .addStatement(
-                                    "final var $L = new $T()",
-                                    "mapper",
-                                    entry.mapper.name)
-                                .addStatement(
-                                    "return $T.defer($L, $L.getName(), $L.getArguments($L))",
-                                    gov.nasa.jpl.aerie.merlin.framework.ModelActions.class,
-                                    "duration",
-                                    "mapper",
-                                    "mapper",
-                                    "activity")
-                                .build(),
-                            MethodSpec
-                                .methodBuilder("defer")
-                                .addModifiers(Modifier.PUBLIC, Modifier.STATIC)
-                                .returns(String.class)
-                                .addParameter(
-                                    ParameterSpec
-                                        .builder(
-                                            TypeName.LONG,
-                                            "quantity")
-                                        .addModifiers(Modifier.FINAL)
-                                        .build())
-                                .addParameter(
-                                    ParameterSpec
-                                        .builder(
-                                            gov.nasa.jpl.aerie.merlin.protocol.types.Duration.class,
-                                            "unit")
-                                        .addModifiers(Modifier.FINAL)
-                                        .build())
-                                .addParameter(
-                                    ClassName.get(entry.declaration),
-                                    "activity",
-                                    Modifier.FINAL)
-                                .addStatement(
-                                    "return defer($L.times($L), $L)",
-                                    "unit",
-                                    "quantity",
-                                    "activity")
-                                .build(),
-                            MethodSpec
-                                .methodBuilder("call")
-                                .addModifiers(Modifier.PUBLIC, Modifier.STATIC)
-                                .returns(TypeName.VOID)
-                                .addParameter(
-                                    ClassName.get(entry.declaration),
-                                    "activity",
-                                    Modifier.FINAL)
-                                .addStatement(
-                                    "$T.waitFor(spawn($L))",
-                                    gov.nasa.jpl.aerie.merlin.framework.ModelActions.class,
-                                    "activity")
-                                .build())
-                        .stream())
-                    .collect(Collectors.toList()))
-            .build();
-
-    return JavaFile
-        .builder(typeName.packageName(), typeSpec)
-        .skipJavaLangImports(true)
-        .build();
-  }
-
-  private JavaFile generateMissionModelFactory(final MissionModelRecord missionModel) {
-    final var typeName = missionModel.getFactoryName();
-
-    final var typeSpec =
-        TypeSpec
-            .classBuilder(typeName)
-            .addAnnotation(
-                AnnotationSpec
-                    .builder(javax.annotation.processing.Generated.class)
-                    .addMember("value", "$S", MissionModelProcessor.class.getCanonicalName())
-                    .build())
-            .addModifiers(Modifier.PUBLIC, Modifier.FINAL)
-            .addSuperinterface(
-                ParameterizedTypeName.get(
-                    ClassName.get(MissionModelFactory.class),
-                    ParameterizedTypeName.get(
-                        ClassName.get(gov.nasa.jpl.aerie.merlin.framework.RootModel.class),
-                        ClassName.get(missionModel.topLevelModel))))
-            .addMethod(
-                MethodSpec
-                    .methodBuilder("getTaskSpecTypes")
-                    .addModifiers(Modifier.PUBLIC)
-                    .addAnnotation(Override.class)
-                    .returns(ParameterizedTypeName.get(
-                        ClassName.get(Map.class),
-                        ClassName.get(String.class),
-                        ParameterizedTypeName.get(
-                            ClassName.get(gov.nasa.jpl.aerie.merlin.protocol.model.TaskSpecType.class),
-                            ParameterizedTypeName.get(
-                                ClassName.get(gov.nasa.jpl.aerie.merlin.framework.RootModel.class),
-                                ClassName.get(missionModel.topLevelModel)),
-                            WildcardTypeName.subtypeOf(Object.class))))
-                    .addStatement("return $T.activityTypes", missionModel.getTypesName())
-                    .build())
-            .addMethod(
-                MethodSpec
-                    .methodBuilder("instantiate")
-                    .addModifiers(Modifier.PUBLIC)
-                    .addAnnotation(Override.class)
-                    .addParameter(
-                        TypeName.get(gov.nasa.jpl.aerie.merlin.protocol.types.SerializedValue.class),
-                        "configuration",
-                        Modifier.FINAL)
-                    .addParameter(
-                        ClassName.get(gov.nasa.jpl.aerie.merlin.protocol.driver.Initializer.class),
-                        "builder",
-                        Modifier.FINAL)
-                    .returns(
-                        ParameterizedTypeName.get(
-                            ClassName.get(gov.nasa.jpl.aerie.merlin.framework.RootModel.class),
-                            ClassName.get(missionModel.topLevelModel)))
-                    .addStatement(
-                        "final var $L = new $T($L)",
-                        "registrar",
-                        gov.nasa.jpl.aerie.merlin.framework.Registrar.class,
-                        "builder")
-                    .addStatement(
-                        "final var $L = $T.makeExecutorService()",
-                        "executor",
-                        gov.nasa.jpl.aerie.merlin.framework.RootModel.class)
-                    .addCode("\n")
-                    .addCode(
-                        missionModel.modelConfiguration
-                            .map(configElem -> CodeBlock  // if configuration is provided
-                                                          .builder()
-                                                          .addStatement(
-                                                              "final var $L = $L",
-                                                              "configMapper",
-                                                              buildConfigurationMapperBlock(missionModel, configElem))
-                                                          .addStatement(
-                                                              "final var $L = $L.deserializeValue($L).getSuccessOrThrow()",
-                                                              "deserializedConfig",
-                                                              "configMapper",
-                                                              "configuration")
-                                                          .addStatement(
-                                                              "final var $L = $T.initializing($L, $L, () -> new $T($L, $L))",
-                                                              "model",
-                                                              gov.nasa.jpl.aerie.merlin.framework.InitializationContext.class,
-                                                              "executor",
-                                                              "builder",
-                                                              ClassName.get(missionModel.topLevelModel),
-                                                              "registrar",
-                                                              "deserializedConfig")
-                                                          .build())
-                            .orElseGet(() -> CodeBlock  // if configuration is not provided
-                                                        .builder()
-                                                        .addStatement(
-                                                            "final var $L = $T.initializing($L, $L, () -> new $T($L))",
-                                                            "model",
-                                                            gov.nasa.jpl.aerie.merlin.framework.InitializationContext.class,
-                                                            "executor",
-                                                            "builder",
-                                                            ClassName.get(missionModel.topLevelModel),
-                                                            "registrar")
-                                                        .build()))
-                    .addCode("\n")
-                    .addStatement(
-                        "return new $T<$T>($L, $L)",
-                        gov.nasa.jpl.aerie.merlin.framework.RootModel.class,
-                        ClassName.get(missionModel.topLevelModel),
-                        "model",
-                        "executor")
-                    .build())
-            .addMethod(
-                MethodSpec
-                    .methodBuilder("getParameters")
-                    .addModifiers(Modifier.PUBLIC)
-                    .addAnnotation(Override.class)
-                    .returns(ParameterizedTypeName.get(List.class, Parameter.class))
-                    .addCode(
-                        missionModel.modelConfiguration
-                            .map(configElem -> CodeBlock  // if configuration is provided
-                                .builder()
-                                .addStatement("return $L.getValueSchema().asStruct().map(parameterMap ->\n"+
-                                              "parameterMap.keySet().stream().map(name -> new $T(name, parameterMap.get(name))).toList())\n"+
-                                              ".orElse($T.of())",
-                                    buildConfigurationMapperBlock(missionModel, configElem),
-                                    Parameter.class,
-                                    List.class)
-                                .build())
-                            .orElse(CodeBlock
-                                .builder()
-                                .addStatement("return $T.of()", List.class)
-                                .build()))
-                    .build())
-            .build();
-
-    return JavaFile
-        .builder(typeName.packageName(), typeSpec)
-        .skipJavaLangImports(true)
-        .build();
-  }
-
-  private JavaFile generateActivityTypes(final MissionModelRecord missionModel) {
-    final var typeName = missionModel.getTypesName();
-
-    final var typeSpec =
-        TypeSpec
-            .classBuilder(typeName)
-            .addAnnotation(
-                AnnotationSpec
-                    .builder(javax.annotation.processing.Generated.class)
-                    .addMember("value", "$S", MissionModelProcessor.class.getCanonicalName())
-                    .build())
-            .addModifiers(Modifier.PUBLIC, Modifier.FINAL)
-            .addField(
-                FieldSpec
-                    .builder(
-                        ParameterizedTypeName.get(
-                            ClassName.get(List.class),
-                            ParameterizedTypeName.get(
-                                ClassName.get(gov.nasa.jpl.aerie.merlin.protocol.model.TaskSpecType.class),
-                                ParameterizedTypeName.get(
-                                    ClassName.get(gov.nasa.jpl.aerie.merlin.framework.RootModel.class),
-                                    ClassName.get(missionModel.topLevelModel)),
-                                WildcardTypeName.subtypeOf(Object.class))),
-                        "activityTypeList",
-                        Modifier.PRIVATE, Modifier.STATIC, Modifier.FINAL)
-                    .initializer(
-                        "$T.of($>$>\n$L$<$<)",
-                        List.class,
-                        missionModel.activityTypes
-                            .stream()
-                            .map(activityType -> CodeBlock.builder().add("new $T()", activityType.mapper.name))
-                            .reduce((x, y) -> x.add(",\n$L", y.build()))
-                            .orElse(CodeBlock.builder())
-                            .build())
-                    .build())
-            .addField(
-                FieldSpec
-                    .builder(
-                        ParameterizedTypeName.get(
-                            ClassName.get(Map.class),
-                            ClassName.get(String.class),
-                            ParameterizedTypeName.get(
-                                ClassName.get(gov.nasa.jpl.aerie.merlin.protocol.model.TaskSpecType.class),
-                                ParameterizedTypeName.get(
-                                    ClassName.get(gov.nasa.jpl.aerie.merlin.framework.RootModel.class),
-                                    ClassName.get(missionModel.topLevelModel)),
-                                WildcardTypeName.subtypeOf(Object.class))),
-                        "activityTypes",
-                        Modifier.PUBLIC, Modifier.STATIC, Modifier.FINAL)
-                    .initializer(
-                        "$L.stream().collect($T.toMap($$ -> $$.getName(), $$ -> $$));",
-                        "activityTypeList",
-                        java.util.stream.Collectors.class)
-                    .build())
-            .build();
-
-    return JavaFile
-        .builder(typeName.packageName(), typeSpec)
-        .skipJavaLangImports(true)
-        .build();
-  }
-
-  private JavaFile generateMerlinPlugin(final MissionModelRecord missionModel) {
-    final var typeName = missionModel.getPluginName();
-
-    final var typeSpec =
-        TypeSpec
-            .classBuilder(typeName)
-            .addAnnotation(
-                AnnotationSpec
-                    .builder(javax.annotation.processing.Generated.class)
-                    .addMember("value", "$S", MissionModelProcessor.class.getCanonicalName())
-                    .build())
-            .addModifiers(Modifier.PUBLIC, Modifier.FINAL)
-            .addSuperinterface(MerlinPlugin.class)
-            .addMethod(
-                MethodSpec
-                    .methodBuilder("getFactory")
-                    .addAnnotation(Override.class)
-                    .addModifiers(Modifier.PUBLIC)
-                    .returns(missionModel.getFactoryName())
-                    .addStatement(
-                        "return new $T()",
-                        missionModel.getFactoryName())
-                    .build())
-            .build();
-
-    return JavaFile
-        .builder(typeName.packageName(), typeSpec)
-        .skipJavaLangImports(true)
-        .build();
   }
 }
