+++ merlin-framework-processor/src/main/java/gov/nasa/jpl/aerie/merlin/processor/generator/ActivityMapperMethodMaker.java
@@ -1,10 +1,10 @@
-package gov.nasa.jpl.aerie.merlin.processor.instantiators;
+package gov.nasa.jpl.aerie.merlin.processor.generator;
 
 import com.squareup.javapoet.CodeBlock;
 import com.squareup.javapoet.MethodSpec;
 import com.squareup.javapoet.ParameterizedTypeName;
 import com.squareup.javapoet.TypeName;
-import gov.nasa.jpl.aerie.merlin.processor.ActivityDefaultsStyle;
+import gov.nasa.jpl.aerie.merlin.processor.metamodel.ActivityDefaultsStyle;
 import gov.nasa.jpl.aerie.merlin.protocol.types.MissingArgumentException;
 import gov.nasa.jpl.aerie.merlin.processor.metamodel.ActivityParameterRecord;
 import gov.nasa.jpl.aerie.merlin.processor.metamodel.ActivityTypeRecord;
@@ -17,7 +17,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
-public interface ActivityMapperInstantiator {
+public interface ActivityMapperMethodMaker {
 
   MethodSpec makeInstantiateMethod(final ActivityTypeRecord activityType);
 
@@ -146,12 +146,12 @@ public interface ActivityMapperInstantiator {
             .build());
   }
 
-  static ActivityMapperInstantiator make(final ActivityDefaultsStyle style) {
+  static ActivityMapperMethodMaker make(final ActivityDefaultsStyle style) {
     return switch (style) {
-      case AllStaticallyDefined -> new AllStaticallyDefinedInstantiator();
-      case NoneDefined -> new NoneDefinedInstantiator();
-      case AllDefined -> new AllDefinedInstantiator();
-      case SomeStaticallyDefined -> new SomeStaticallyDefinedInstantiator();
+      case AllStaticallyDefined -> new AllStaticallyDefinedMethodMaker();
+      case NoneDefined -> new NoneDefinedMethodMaker();
+      case AllDefined -> new AllDefinedMethodMaker();
+      case SomeStaticallyDefined -> new SomeStaticallyDefinedMethodMaker();
     };
   }
 }
