+++ merlin-framework-processor/src/main/java/gov/nasa/jpl/aerie/merlin/processor/generator/AllStaticallyDefinedMethodMaker.java
@@ -1,4 +1,4 @@
-package gov.nasa.jpl.aerie.merlin.processor.instantiators;
+package gov.nasa.jpl.aerie.merlin.processor.generator;
 
 import com.squareup.javapoet.ClassName;
 import com.squareup.javapoet.CodeBlock;
@@ -15,7 +15,7 @@ import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
-public class AllStaticallyDefinedInstantiator implements ActivityMapperInstantiator {
+public class AllStaticallyDefinedMethodMaker implements ActivityMapperMethodMaker {
 
   @Override
   public MethodSpec makeInstantiateMethod(final ActivityTypeRecord activityType) {
@@ -93,7 +93,7 @@ public class AllStaticallyDefinedInstantiator implements ActivityMapperInstantia
       break;
     }
 
-    methodBuilder = ActivityMapperInstantiator
+    methodBuilder = ActivityMapperMethodMaker
         .makeArgumentPresentCheck(methodBuilder, activityType).addCode("\n");
 
     // Add return statement with instantiation of class with parameters
