+++ merlin-framework-processor/src/main/java/gov/nasa/jpl/aerie/merlin/processor/generator/NoneDefinedMethodMaker.java
@@ -1,4 +1,4 @@
-package gov.nasa.jpl.aerie.merlin.processor.instantiators;
+package gov.nasa.jpl.aerie.merlin.processor.generator;
 
 import com.squareup.javapoet.ClassName;
 import com.squareup.javapoet.CodeBlock;
@@ -13,7 +13,7 @@ import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
-public class NoneDefinedInstantiator implements ActivityMapperInstantiator {
+public class NoneDefinedMethodMaker implements ActivityMapperMethodMaker {
 
   @Override
   public MethodSpec makeInstantiateMethod(final ActivityTypeRecord activityType) {
@@ -80,7 +80,7 @@ public class NoneDefinedInstantiator implements ActivityMapperInstantiator {
         .endControlFlow()
         .endControlFlow().addCode("\n");
 
-    methodBuilder = ActivityMapperInstantiator
+    methodBuilder = ActivityMapperMethodMaker
         .makeArgumentPresentCheck(methodBuilder, activityType).addCode("\n");
 
     // Add return statement with instantiation of class with parameters
