+++ merlin-framework-processor/src/main/java/gov/nasa/jpl/aerie/merlin/processor/metamodel/ActivityTypeRecord.java
@@ -7,7 +7,6 @@ import javax.lang.model.element.TypeElement;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
-import gov.nasa.jpl.aerie.merlin.processor.ActivityDefaultsStyle;
 
 public final class ActivityTypeRecord {
   public final TypeElement declaration;
